requirements-prd.txt
txt
Django==3.1.3
pymysql
djangorestframework==3.12.2
django-filter
cryptography
django-notifications-hq
django-debug-toolbar
django-storages
boto3
pymemcache
redis==4.5.4
celery
django-ratelimit
